requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.22
python-dotenv==0.18.0